% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamaparse.R
\name{check_status}
\alias{check_status}
\title{Check status of conversion}
\usage{
check_status(
  job_id = "e778beda-74bd-4f55-a700-beafed677556",
  api_key = Sys.getenv("LLAMA_CLOUD_API_KEY")
)
}
\description{
Check status of conversion
}
\keyword{internal}
