% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_data.R
\name{chunk_data}
\alias{chunk_data}
\title{Chunk the markdown, combine with metadata and generate id.}
\usage{
chunk_data(markdown, meta)
}
\description{
Chunk the markdown, combine with metadata and generate id.
}
\keyword{internal}
