% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamaparse.R
\name{get_markdown}
\alias{get_markdown}
\title{Fetch processed markdown from server}
\usage{
get_markdown(
  job_id = "e778beda-74bd-4f55-a700-beafed677556",
  api_key = "llx-XTMPRKlsgx8whmOpf4nKf5m1X4EgNnAvJprNjHfoXydNiSLY"
)
}
\description{
Fetch processed markdown from server
}
\keyword{internal}
