% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamaparse.R
\name{llamaparse}
\alias{llamaparse}
\title{Convert PDF to Markdown using Llamaparse.}
\usage{
llamaparse(
  filepath = system.file("extdata/realist-evaluation.pdf", package =
    "llamaparse2chromadb"),
  api_key = Sys.getenv("LLAMA_CLOUD_API_KEY"),
  timeout = 60
)
}
\description{
Convert PDF to Markdown using Llamaparse.
}
\keyword{internal}
