% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2chromadb.R
\name{pdf_to_chromadb}
\alias{pdf_to_chromadb}
\title{PDF to Chromadb}
\usage{
pdf_to_chromadb(
  api_key = Sys.getenv("LLAMA_CLOUD_API_KEY"),
  open_ai_api_key = Sys.getenv("OPENAI_API_KEY"),
  path_to_chromadb = Sys.getenv("PATH_TO_CHROMADB"),
  default_collection = Sys.getenv("CHROMADB_DEFAULT_COLLECTION")
)
}
\arguments{
\item{api_key}{Llamacloud api key.}

\item{open_ai_api_key}{Open Ai api key}

\item{path_to_chromadb}{Path to chromadb}

\item{default_collection}{Default collection in chromadb}
}
\value{
A list.
}
\description{
RStudio addin to add a pdf to chromadb vector database. The function
}
\examples{
\dontrun{
pdf_to_chromadb()
}
}
