% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_data.R
\name{semantic_split}
\alias{semantic_split}
\title{Function to split text based on embedding threshold.}
\usage{
semantic_split(x)
}
\description{
Function to split text based on embedding threshold.
}
\keyword{internal}
